feature/adding project structure
