+++ src/styles/Home.css
@@ -224,7 +224,7 @@ body {
     overflow: hidden;
 }
 html {
-    background: url('./bg.png');
+    background: url('../bg.png');
     background-size: cover;
     background-attachment: fixed;
     background-position: 60% center;
